Renaming from aiq_simple to aiq_simple_web_query
