+++ examples/src/simple_web_query/pyproject.toml
@@ -6,7 +6,7 @@ requires = ["setuptools >= 64", "setuptools-scm>=8"]
 root = "../../../.."
 
 [project]
-name = "aiq_simple"
+name = "aiq_simple_web_query"
 dynamic = ["version"]
 dependencies = [
   "aiqtoolkit[langchain]~=1.2",
@@ -21,4 +21,4 @@ classifiers = ["Programming Language :: Python"]
 aiqtoolkit = { path = "../../../..", editable = true }
 
 [project.entry-points.'aiq.components']
-aiq_simple = "aiq_simple.register"
+aiq_simple_web_query = "aiq_simple_web_query.register"
