+++ examples/src/simple_web_query/src/aiq_simple_web_query/configs/config.yml
