+++ examples/src/simple_web_query/tests/test_simple_workflow.py
@@ -20,7 +20,7 @@ import logging
 from pathlib import Path
 
 import pytest
-from aiq_simple.register import WebQueryToolConfig
+from aiq_simple_web_query.register import WebQueryToolConfig
 
 from aiq.runtime.loader import load_workflow
