+++ examples/src/simple_web_query_eval/pyproject.toml
@@ -13,7 +13,7 @@ name = "aiq_simple_eval"
 dynamic = ["version"]
 dependencies = [
   "aiqtoolkit[langchain]~=1.2",
-  "aiq_simple",
+  "aiq_simple_web_query",
 ]
 requires-python = ">=3.11,<3.13"
 description = "Simple LangSmith-Documentation Agent Evaluation and Profiling - demonstrates AIQ toolkit evaluation capabilities"
@@ -22,4 +22,4 @@ classifiers = ["Programming Language :: Python"]
 
 [tool.uv.sources]
 aiqtoolkit = { path = "../../../..", editable = true }
-aiq_simple = { path = "../../../basic/functions/simple", editable = true }
+aiq_simple_web_query = { path = "../../../basic/functions/simple", editable = true }
